Added Team Generation functionality
Added a Character component and imported it into the project. The components data function returns an empty object and it also has the fetch method from the app and a switch character method that randomizes the fetch URL character number. Added created function so it fired upon creation. Inside the project there is now an initial ids array and is binded to the character via the props array added on the character component. Should randomize ids within cards. Added switch character to cards on click. Styled cards a bit.